fix reconnection bug; add some new APIs; change scan mode when screen is off

+++ easyble/src/main/java/com/ficat/easyble/scan/BleScan.java
@@ -3,7 +3,7 @@ package com.ficat.easyble.scan;
 import java.util.UUID;
 
 public interface BleScan<T> {
-    void startScan(int scanPeriod, String scanDeviceName, String scanDeviceAddress, UUID[] scanServiceUuids, T callback);
+    void startScan(long scanPeriod, String scanDeviceName, String scanDeviceAddress, UUID[] scanServiceUuids, T callback);
 
     void stopScan();
